error handling for extracting from json object
the pfx_event json object sometimes does not include certain fields
(sometimes null). we previously assumes data is correct and abused the
usage of `unwrap` too much. now it handles the missing cases and
returns empty lists for attacker and victims if this happens.

this should addresses issue caida/bgphijacks-tools#306